improve

+++ database/migrations/2021_06_08_162727_create_tel_users_table.php
@@ -14,8 +14,8 @@ class CreateTelUsersTable extends Migration
     public function up()
     {
         Schema::create('tel_users', function (Blueprint $table) {
-            $table->integer('user_id');
-            $table->integer('chat_id');
+            $table->bigInteger('user_id');
+            $table->bigInteger('chat_id');
             $table->string('first_name');
             $table->string('last_name')->nullable();
             $table->string('username')->nullable();
